+++ src/internal.ts
@@ -1,9 +1,9 @@
-export * from "./crap/env"
-export * from "./crap/errors"
+export * from "./utils/env"
+export * from "./utils/errors"
 export * from "./types/types-external"
 export * from "./types/types-internal"
-export * from "./crap/common"
-export * from "./crap/plugins"
+export * from "./utils/common"
+export * from "./utils/plugins"
 export * from "./core/scope"
 export * from "./core/finalize"
 export * from "./core/proxy"
